@@ -1,98 +1,101 @@
 {
-    "name": "twitch-desktop",
-    "version": "2.5.0",
-    "description": "Twitch Desktop Client",
-    "homepage": "https://github.com/hzeroo/twitch-desktop",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/hzeroo/twitch-desktop.git"
-    },
-    "author": {
-        "name": "hzeroo",
-        "email": "edusvh@gmail.com"
-    },
-    "keywords": [
-        "twitch",
-        "desktop"
-    ],
-    "main": "main.js",
-    "private": true,
-    "scripts": {
-        "postinstall": "yarn postinstall:electron && electron-builder install-app-deps",
-        "postinstall:web": "node postinstall-web",
-        "postinstall:electron": "node postinstall",
-        "ng": "ng",
-        "start": "yarn postinstall:electron && npm-run-all -p ng:serve electron:serve",
-        "build": "yarn postinstall:electron && yarn electron:serve-tsc && ng build",
-        "build:dev": "yarn build -c dev",
-        "build:prod": "yarn build -c production",
-        "ng:serve": "ng serve",
-        "ng:serve:web": "yarn postinstall:web && ng serve -o",
-        "electron:serve-tsc": "tsc -p tsconfig-serve.json",
-        "electron:serve": "wait-on http-get://localhost:4200/ && yarn electron:serve-tsc && electron . --serve",
-        "release:linux": "yarn build:prod && electron-builder build --linux",
-        "release:windows": "yarn build:prod && electron-builder build --windows",
-        "release:mac": "yarn build:prod && electron-builder build --mac",
-        "release:all": "yarn build:prod && electron-builder build --windows --linux --mac",
-        "publish:linux": "yarn build:prod && electron-builder build --linux -p always",
-        "publish:windows": "yarn build:prod && electron-builder build --windows -p always",
-        "publish:all": "yarn build:prod && electron-builder build --windows --linux --mac -p always"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "0.803.6",
-        "@angular/animations": "^8.2.8",
-        "@angular/cli": "8.3.6",
-        "@angular/common": "8.2.8",
-        "@angular/compiler": "8.2.8",
-        "@angular/compiler-cli": "8.2.8",
-        "@angular/core": "8.2.8",
-        "@angular/forms": "8.2.8",
-        "@angular/http": "7.2.15",
-        "@angular/language-service": "8.2.8",
-        "@angular/platform-browser": "8.2.8",
-        "@angular/platform-browser-dynamic": "8.2.8",
-        "@angular/router": "8.2.8",
-        "@ng-bootstrap/ng-bootstrap": "^5.1.1",
-        "@ngx-translate/core": "11.0.1",
-        "@ngx-translate/http-loader": "4.0.0",
-        "@types/lodash": "^4.14.139",
-        "@types/node": "12.7.8",
-        "@types/request-promise-native": "^1.0.17",
-        "apollo-angular": "^1.7.0",
-        "apollo-angular-link-http": "^1.8.0",
-        "apollo-cache-inmemory": "^1.6.3",
-        "apollo-client": "^2.6.4",
-        "apollo-link": "^1.2.13",
-        "apollo-link-retry": "^2.2.15",
-        "bootstrap-scss": "^4.3.1",
-        "clappr": "^0.3.9",
-        "core-js": "3.2.1",
-        "electron": "6.0.10",
-        "electron-builder": "21.2.0",
-        "electron-reload": "1.5.0",
-        "electron-store": "^5.0.0",
-        "graphql": "^14.5.8",
-        "graphql-tag": "^2.10.1",
-        "level-selector": "^0.2.0",
-        "lodash": "^4.17.15",
-        "material-icons": "^0.3.1",
-        "moment": "^2.24.0",
-        "normalize-scss": "^7.0.1",
-        "npm-run-all": "4.1.5",
-        "roboto-fontface": "^0.10.0",
-        "rxjs": "6.5.3",
-        "simplebar": "^4.2.3",
-        "ts-node": "8.4.1",
-        "tslint": "5.20.0",
-        "typescript": "3.5.3",
-        "wait-on": "3.3.0",
-        "webdriver-manager": "13.0.0",
-        "zone.js": "~0.9.1"
-    },
-    "dependencies": {
-        "electron-log": "^3.0.8",
-        "electron-updater": "^4.1.2",
-        "request": "^2.88.0",
-        "request-promise-native": "^1.0.7"
-    }
+  "name": "twitch-desktop",
+  "version": "2.6.0",
+  "description": "Twitch Desktop Client",
+  "homepage": "https://github.com/hzeroo/twitch-desktop",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/hzeroo/twitch-desktop.git"
+  },
+  "author": {
+    "name": "hzeroo",
+    "email": "edusvh@gmail.com"
+  },
+  "keywords": [
+    "twitch",
+    "desktop"
+  ],
+  "main": "main.js",
+  "private": true,
+  "scripts": {
+    "postinstall": "yarn postinstall:electron && electron-builder install-app-deps",
+    "postinstall:web": "node postinstall-web",
+    "postinstall:electron": "node postinstall",
+    "ng": "ng",
+    "start": "yarn postinstall:electron && npm-run-all -p ng:serve electron:serve",
+    "build": "yarn postinstall:electron && yarn electron:serve-tsc && ng build",
+    "build:dev": "yarn build -c dev",
+    "build:prod": "yarn build -c production",
+    "ng:serve": "ng serve",
+    "ng:serve:web": "yarn postinstall:web && ng serve -o",
+    "electron:serve-tsc": "tsc -p tsconfig-serve.json",
+    "electron:serve": "wait-on http-get://localhost:4200/ && yarn electron:serve-tsc && electron . --serve",
+    "release:linux": "yarn build:prod && electron-builder build --linux",
+    "release:windows": "yarn build:prod && electron-builder build --windows",
+    "release:mac": "yarn build:prod && electron-builder build --mac",
+    "release:all": "yarn build:prod && electron-builder build --windows --linux --mac",
+    "publish:linux": "yarn build:prod && electron-builder build --linux -p always",
+    "publish:windows": "yarn build:prod && electron-builder build --windows -p always",
+    "publish:all": "yarn build:prod && electron-builder build --windows --linux --mac -p always"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular/animations": "^9.0.0",
+    "@angular/cli": "9.0.1",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/compiler-cli": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/forms": "9.0.0",
+    "@angular/language-service": "9.0.0",
+    "@angular/localize": "^9.0.0",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.1",
+    "@ngx-translate/core": "12.0.0",
+    "@ngx-translate/http-loader": "4.0.0",
+    "@types/lodash": "^4.14.139",
+    "@types/node": "12.12.6",
+    "@types/request-promise-native": "^1.0.17",
+    "apollo-angular": "^1.7.0",
+    "apollo-angular-link-http": "^1.8.0",
+    "apollo-cache-inmemory": "^1.6.3",
+    "apollo-client": "^2.6.4",
+    "apollo-link": "^1.2.13",
+    "apollo-link-retry": "^2.2.15",
+    "bootstrap-scss": "^4.3.1",
+    "clappr": "^0.3.9",
+    "core-js": "3.6.4",
+    "electron": "8.0.0",
+    "electron-builder": "22.3.2",
+    "electron-reload": "1.5.0",
+    "electron-store": "^5.0.0",
+    "graphql": "^14.5.8",
+    "graphql-tag": "^2.10.1",
+    "level-selector": "^0.2.0",
+    "lodash": "^4.17.15",
+    "material-icons": "^0.3.1",
+    "moment": "^2.24.0",
+    "normalize-scss": "^7.0.1",
+    "npm-run-all": "4.1.5",
+    "prettier": "^1.19.1",
+    "roboto-fontface": "^0.10.0",
+    "rxjs": "6.5.4",
+    "simplebar": "^5.1.0",
+    "ts-node": "8.6.2",
+    "tslint": "6.0.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "3.7.5",
+    "wait-on": "4.0.0",
+    "webdriver-manager": "13.0.0",
+    "zone.js": "~0.10.2"
+  },
+  "dependencies": {
+    "electron-log": "^4.0.6",
+    "electron-updater": "^4.1.2",
+    "request": "^2.88.0",
+    "request-promise-native": "^1.0.7",
+    "tslib": "^1.10.0"
+  }
 }
