@@ -1,22 +1,22 @@
-import "reflect-metadata";
-import "../polyfills";
+import 'reflect-metadata';
+import '../polyfills';
 
-import { BrowserModule } from "@angular/platform-browser";
-import { NgModule, Pipe, PipeTransform } from "@angular/core";
-import { FormsModule } from "@angular/forms";
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
-import moment from "moment";
-import { HttpClientModule, HttpClient } from "@angular/common/http";
-import { AppRoutingModule } from "./app-routing.module";
-import { TranslateModule, TranslateLoader } from "@ngx-translate/core";
-import { TranslateHttpLoader } from "@ngx-translate/http-loader";
-import { GraphQLModule } from "./graphql.module";
+import {HttpClient, HttpClientModule} from '@angular/common/http';
+import {NgModule, Pipe, PipeTransform} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {BrowserModule} from '@angular/platform-browser';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
+import {TranslateHttpLoader} from '@ngx-translate/http-loader';
+import moment from 'moment';
+import {AppRoutingModule} from './app-routing.module';
+import {GraphQLModule} from './graphql.module';
 
 // Pipes
 // Transforms a time string into a string representing time since that moment
 @Pipe({
-  name: "twTimeSince"
+  name: 'twTimeSince'
 })
 export class TimeSincePipe implements PipeTransform {
   transform(startTime: string) {
@@ -25,17 +25,12 @@ export class TimeSincePipe implements PipeTransform {
 }
 
 @Pipe({
-  name: "twScaleImage"
+  name: 'twScaleImage'
 })
 export class ScaleImagePipe implements PipeTransform {
   transform(url: string, width: string, height: string) {
-    if (
-      url &&
-      url !== "" &&
-      url.includes("{width}") &&
-      url.includes("{height}")
-    ) {
-      return url.replace("{width}", width).replace("{height}", height);
+    if (url && url !== '' && url.includes('{width}') && url.includes('{height}')) {
+      return url.replace('{width}', width).replace('{height}', height);
     } else {
       return url;
     }
@@ -43,35 +38,36 @@ export class ScaleImagePipe implements PipeTransform {
 }
 
 // Components
-import { AppComponent } from "./app.component";
-import { PlayerComponent } from "./components/player/player.component";
-import { GamesComponent } from "./components/games/games.component";
-import { ChannelsComponent } from "./components/channels/channels.component";
-import { ToolbarComponent } from "./components/toolbar/toolbar.component";
-import { LoginComponent } from "./components/login/login.component";
-import { SpinnerComponent } from "./components/spinner/spinner.component";
-import { SidebarComponent } from "./components/sidebar/sidebar.component";
-import { ErrorComponent } from "./components/error-handler/errorhandler.component";
-import { ChatComponent } from "./components/player/chat/chat.component";
-import { SettingsComponent } from "./components/settings/settings.component";
+import {AppComponent} from './app.component';
+import {ChannelsComponent} from './components/channels/channels.component';
+import {ErrorComponent} from './components/error-handler/errorhandler.component';
+import {GamesComponent} from './components/games/games.component';
+import {LoginFormComponent} from './components/login/login-form/login-form.component';
+import {LoginComponent} from './components/login/login.component';
+import {ChatComponent} from './components/player/chat/chat.component';
+import {PlayerComponent} from './components/player/player.component';
+import {SettingsComponent} from './components/settings/settings.component';
+import {SidebarComponent} from './components/sidebar/sidebar.component';
+import {SpinnerComponent} from './components/spinner/spinner.component';
+import {ToolbarComponent} from './components/toolbar/toolbar.component';
 
 // Services
-import { ElectronService } from "./providers/electron.service";
-import { ErrorService } from "./providers/errorhandler.service";
-import { SpinnerService } from "./providers/spinner.service";
-import { TwitchService } from "./providers/twitch.service";
-import { ToolbarService } from "./providers/toolbar.service";
-import { GameService } from "./providers/games.service";
-import { ChannelService } from "./providers/channels.service";
-import { SettingsService } from "./providers/settings.service";
+import {ChannelService} from './providers/channels.service';
+import {ElectronService} from './providers/electron.service';
+import {ErrorService} from './providers/errorhandler.service';
+import {GameService} from './providers/games.service';
+import {SettingsService} from './providers/settings.service';
+import {SpinnerService} from './providers/spinner.service';
+import {ToolbarService} from './providers/toolbar.service';
+import {TwitchService} from './providers/twitch.service';
 
 // Directives
-import { WebviewDirective } from "./directives/webview.directive";
-import { InfiniteScroll } from "./directives/infinitescroll.directive";
+import {InfiniteScroll} from './directives/infinitescroll.directive';
+import {WebviewDirective} from './directives/webview.directive';
 
 // AoT requires an exported function for factories
 export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http, "./assets/i18n/", ".json");
+  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
 }
 
 @NgModule({
@@ -88,6 +84,7 @@ export function HttpLoaderFactory(http: HttpClient) {
     ErrorComponent,
     ChatComponent,
     SettingsComponent,
+    LoginFormComponent,
 
     // Directives
     WebviewDirective,
@@ -130,4 +127,4 @@ export function HttpLoaderFactory(http: HttpClient) {
 
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
