@@ -1,13 +1,13 @@
-import { Component, OnInit } from "@angular/core";
-import { Router } from "@angular/router";
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
-import { SpinnerService } from "../../providers/spinner.service";
+import { SpinnerService } from '../../providers/spinner.service';
 
 // Spinner Component
 @Component({
-  templateUrl: "./spinner.component.html",
-  selector: "tw-spinner",
-  styleUrls: ["./spinner.component.scss"]
+  templateUrl: './spinner.component.html',
+  selector: 'tw-spinner',
+  styleUrls: ['./spinner.component.scss']
 })
 export class SpinnerComponent implements OnInit {
   isLoading = false;
@@ -19,5 +19,5 @@ export class SpinnerComponent implements OnInit {
     });
   }
 
-  ngOnInit() { }
+  ngOnInit() {}
 }
