@@ -1,24 +1,29 @@
-import { NgModule } from "@angular/core";
-import { ApolloModule, APOLLO_OPTIONS } from "apollo-angular";
-import { ApolloLink } from "apollo-link";
-import { HttpLinkModule, HttpLink } from "apollo-angular-link-http";
-import { RetryLink } from "apollo-link-retry";
-import { InMemoryCache } from "apollo-cache-inmemory";
-import { HttpHeaders } from "@angular/common/http";
+import { HttpHeaders } from '@angular/common/http';
+import { NgModule } from '@angular/core';
+import { APOLLO_OPTIONS, ApolloModule } from 'apollo-angular';
+import { HttpLink, HttpLinkModule } from 'apollo-angular-link-http';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { ApolloLink } from 'apollo-link';
+import { RetryLink } from 'apollo-link-retry';
 
-const uri = "https://gql.twitch.tv/gql";
+const uri = 'https://gql.twitch.tv/gql';
 export function createApollo(httpLink: HttpLink) {
-
   // Middleware to inject Client-ID header on all GraphQL requests
   const authMiddleware = new ApolloLink((operation, forward) => {
     if (localStorage.getItem('auth_token')) {
       operation.setContext({
-        headers: new HttpHeaders().set("Authorization", `OAuth ${localStorage.getItem('auth_token')}`)
+        headers: new HttpHeaders().set(
+          'Authorization',
+          `OAuth ${localStorage.getItem('auth_token')}`
+        )
       });
       return forward(operation);
     } else {
       operation.setContext({
-        headers: new HttpHeaders().set("Client-ID", "kimne78kx3ncx6brgo4mv6wki5h1ko")
+        headers: new HttpHeaders().set(
+          'Client-ID',
+          'kimne78kx3ncx6brgo4mv6wki5h1ko'
+        )
       });
       return forward(operation);
     }
@@ -27,13 +32,13 @@ export function createApollo(httpLink: HttpLink) {
   const defaultOptions = {
     watchQuery: {
       fetchPolicy: 'network-only',
-      errorPolicy: 'ignore',
+      errorPolicy: 'ignore'
     },
     query: {
       fetchPolicy: 'network-only',
-      errorPolicy: 'all',
-    },
-  }
+      errorPolicy: 'all'
+    }
+  };
 
   return {
     link: ApolloLink.from([
@@ -42,7 +47,7 @@ export function createApollo(httpLink: HttpLink) {
       httpLink.create({ uri })
     ]),
     cache: new InMemoryCache(),
-    defaultOptions: defaultOptions
+    defaultOptions
   };
 }
 
@@ -56,4 +61,4 @@ export function createApollo(httpLink: HttpLink) {
     }
   ]
 })
-export class GraphQLModule { }
+export class GraphQLModule {}
