@@ -1,18 +1,18 @@
-import { Injectable } from "@angular/core";
-import { ApolloQueryResult } from "apollo-client";
-import { TwitchService } from "./twitch.service";
+import { Injectable } from '@angular/core';
+import { ApolloQueryResult } from 'apollo-client';
+import { concat as _concat, map as _map, uniqBy as _uniqBy } from 'lodash';
+import { IGame } from './games.service';
 import {
+  GetCurrentUserOnlineFollowsGQL,
   GetGameStreamsGQL,
   GetTopStreamsGQL,
-  GetCurrentUserOnlineFollowsGQL,
-  TopStreamsResponse,
-  GameStreamsResponse,
-  FollowsResponse
-} from "./twitch-graphql.service";
-import { Game } from "./games.service";
-import { map as _map, concat as _concat, uniqBy as _uniqBy } from "lodash";
+  IFollowsResponse,
+  IGameStreamsResponse,
+  ITopStreamsResponse
+} from './twitch-graphql.service';
+import { TwitchService } from './twitch.service';
 
-export interface Stream {
+export interface IStream {
   id: string;
   broadcaster: {
     id: string;
@@ -23,9 +23,8 @@ export interface Stream {
       game: {
         id: string;
         name: string;
-      }
-    }
-
+      };
+    };
   };
   previewImageURL: string;
   viewersCount: number;
@@ -40,76 +39,89 @@ enum StreamListType {
 // Service that allows components to get channel list information
 @Injectable()
 export class ChannelService {
-  private streams: Stream[] = [];
-  private active_stream: Stream;
+  private streams: IStream[] = [];
+  private active_stream: IStream;
   private streams_list_type: StreamListType;
-  private cursor = "";
-  private game: Game;
+  private cursor = '';
+  private game: IGame;
 
-  constructor(private twitchService: TwitchService,
+  constructor(
+    private twitchService: TwitchService,
     private getGameStreamsGQL: GetGameStreamsGQL,
     private getTopStreamsGQL: GetTopStreamsGQL,
-    private getOnlineFollowsGQL: GetCurrentUserOnlineFollowsGQL) { }
+    private getOnlineFollowsGQL: GetCurrentUserOnlineFollowsGQL
+  ) {}
 
-  set currentStream(stream: Stream) {
+  set currentStream(stream: IStream) {
     this.active_stream = stream;
   }
 
-  get currentStream(): Stream {
+  get currentStream(): IStream {
     return this.active_stream;
   }
 
   getTopStreams() {
     return new Promise((resolve, reject) => {
-      this.getTopStreamsGQL.fetch().subscribe((result: ApolloQueryResult<TopStreamsResponse>) => {
-        if (result.data) {
-          this.streams_list_type = StreamListType.TopStreams;
-          this.cursor = result.data.streams.edges.slice(-1)[0].cursor;
-          this.streams = _map(result.data.streams.edges, (e) => {
-            return e.node;
-          });
+      this.getTopStreamsGQL
+        .fetch()
+        .subscribe((result: ApolloQueryResult<ITopStreamsResponse>) => {
+          if (result.data) {
+            this.streams_list_type = StreamListType.TopStreams;
+            this.cursor = result.data.streams.edges.slice(-1)[0].cursor;
+            this.streams = _map(result.data.streams.edges, e => {
+              return e.node;
+            });
 
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 
-  getGameStreams(game: Game) {
+  getGameStreams(game: IGame) {
     return new Promise((resolve, reject) => {
-      this.getGameStreamsGQL.fetch({ name: game.name }).subscribe((result: ApolloQueryResult<GameStreamsResponse>) => {
-        if (result.data) {
-          this.game = game;
-          this.streams_list_type = StreamListType.GameStreams;
-          this.cursor = result.data.game.streams.edges.slice(-1)[0].cursor;
-          this.streams = _map(result.data.game.streams.edges, (e) => {
-            return e.node;
-          });
+      this.getGameStreamsGQL
+        .fetch({ name: game.name })
+        .subscribe((result: ApolloQueryResult<IGameStreamsResponse>) => {
+          if (result.data) {
+            this.game = game;
+            this.streams_list_type = StreamListType.GameStreams;
+            this.cursor = result.data.game.streams.edges.slice(-1)[0].cursor;
+            this.streams = _map(result.data.game.streams.edges, e => {
+              return e.node;
+            });
 
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 
   getFollowedStreams() {
     return new Promise((resolve, reject) => {
-      this.getOnlineFollowsGQL.fetch().subscribe((result: ApolloQueryResult<FollowsResponse>) => {
-        if (result.data) {
-          this.streams_list_type = StreamListType.FollowingStreams;
-          this.cursor = result.data.currentUser.followedLiveUsers.edges.slice(-1)[0].cursor;
-          this.streams = _map(result.data.currentUser.followedLiveUsers.edges, (e) => {
-            return e.node.stream;
-          });
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+      this.getOnlineFollowsGQL
+        .fetch()
+        .subscribe((result: ApolloQueryResult<IFollowsResponse>) => {
+          if (result.data) {
+            this.streams_list_type = StreamListType.FollowingStreams;
+            this.cursor = result.data.currentUser.followedLiveUsers.edges.slice(
+              -1
+            )[0].cursor;
+            this.streams = _map(
+              result.data.currentUser.followedLiveUsers.edges,
+              e => {
+                return e.node.stream;
+              }
+            );
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 
@@ -131,64 +143,75 @@ export class ChannelService {
 
   private fetchMoreTopStreams() {
     return new Promise((resolve, reject) => {
-      this.getTopStreamsGQL.fetch({ cursor: this.cursor }).subscribe((result: ApolloQueryResult<TopStreamsResponse>) => {
-        if (result.data) {
-          this.cursor = result.data.streams.edges.slice(-1)[0].cursor;
-          this.streams = _uniqBy(
-            _concat(
-              this.streams,
-              _map(result.data.streams.edges, (e) => e.node)
-            ),
-            "id"
-          );
+      this.getTopStreamsGQL
+        .fetch({ cursor: this.cursor })
+        .subscribe((result: ApolloQueryResult<ITopStreamsResponse>) => {
+          if (result.data) {
+            this.cursor = result.data.streams.edges.slice(-1)[0].cursor;
+            this.streams = _uniqBy(
+              _concat(
+                this.streams,
+                _map(result.data.streams.edges, e => e.node)
+              ),
+              'id'
+            );
 
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 
-  private fetchMoreGameStreams(game: Game) {
+  private fetchMoreGameStreams(game: IGame) {
     return new Promise((resolve, reject) => {
-      this.getGameStreamsGQL.fetch({ name: game.name, cursor: this.cursor }).subscribe((result: ApolloQueryResult<GameStreamsResponse>) => {
-        if (result.data) {
-          this.cursor = result.data.game.streams.edges.slice(-1)[0].cursor;
-          this.streams = _uniqBy(
-            _concat(
-              this.streams,
-              _map(result.data.game.streams.edges, (e) => e.node)
-            ),
-            "id"
-          );
+      this.getGameStreamsGQL
+        .fetch({ name: game.name, cursor: this.cursor })
+        .subscribe((result: ApolloQueryResult<IGameStreamsResponse>) => {
+          if (result.data) {
+            this.cursor = result.data.game.streams.edges.slice(-1)[0].cursor;
+            this.streams = _uniqBy(
+              _concat(
+                this.streams,
+                _map(result.data.game.streams.edges, e => e.node)
+              ),
+              'id'
+            );
 
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 
   private fetchMoreFollowedStreams() {
     return new Promise((resolve, reject) => {
-      this.getOnlineFollowsGQL.fetch({ cursor: this.cursor }).subscribe((result: ApolloQueryResult<FollowsResponse>) => {
-        if (result.data) {
-          this.cursor = result.data.currentUser.followedLiveUsers.edges.slice(-1)[0].cursor;
-          this.streams = _uniqBy(
-            _concat(
-              this.streams,
-              _map(result.data.currentUser.followedLiveUsers.edges, (e) => e.node.stream)
-            ),
-            "id"
-          );
+      this.getOnlineFollowsGQL
+        .fetch({ cursor: this.cursor })
+        .subscribe((result: ApolloQueryResult<IFollowsResponse>) => {
+          if (result.data) {
+            this.cursor = result.data.currentUser.followedLiveUsers.edges.slice(
+              -1
+            )[0].cursor;
+            this.streams = _uniqBy(
+              _concat(
+                this.streams,
+                _map(
+                  result.data.currentUser.followedLiveUsers.edges,
+                  e => e.node.stream
+                )
+              ),
+              'id'
+            );
 
-          resolve(this.streams);
-        } else {
-          reject();
-        }
-      });
+            resolve(this.streams);
+          } else {
+            reject();
+          }
+        });
     });
   }
 }
